Revert "Fixed the covariance matrix bug by replacing -a.tau with -1 (both the main class and the ghost curve class need to be fixed)"
This reverts commit 1206ad6ed18b58237e417a17c1aa9756e891f7de.
Reverting this commit because we changed the variance covariance matrix
when it was actually correct (everything else in this branch will remain
the same). This specific commit was reverted in the event the user
wanted to download this version for whatever reason. We don't want them
to potentially get inaccurate results